chore(EMS-1963): rename map CIS countries directory
