+++ src/api/helpers/map-CIS-countries/map-CIS-country/can-apply-offline/index.ts
