+++ src/api/helpers/map-CIS-countries/map-CIS-country/map-risk-category/index.test.ts
