+++ src/api/helpers/map-CIS-countries/map-CIS-country/map-short-term-cover-available/index.ts
